Reports - Added update_all, removed pres.
