+++ models/electronic_new/__init__.py
@@ -10,7 +10,5 @@ from . import electronic_order
 from . import electronic_container
 
 # Account
-from . import account_line_pre
 from . import account_line
-from . import account_contasis_pre
 from . import account_contasis
