+++ views/electronic/account_contasis.xml
@@ -1,9 +1,6 @@
 <?xml version="1.0"?>
 <openerp>
     <data>
-        
-
-
 <!-- FORM --> 
         <record model="ir.ui.view" id="account_contasis_form_view">
             <field name="name">account.contasis.form</field>
@@ -14,27 +11,21 @@
                     <!-- Header -->
                     <header>
 
+                      <!-- Update all --> 
+          						<button 
+          							type="object" 
+          							class="oe_highlight"
+          							name="update" 
+          							string="UPDATE"
+          						/>
 
-<!--
-                                attrs="{'invisible':[   
-                                                        ('x_type','not in',['order']), 
-                                            ]}"
--->
                         <!-- Orders -->
                         <button 
                                 type="object" 
                                 class="oe_highlight"
-
-                                name="update"
-
+                                name="update_account"
                                 string="2019 - Actualizar Ventas"                              
                         />
-
-
-
-<!--
-                                name="%(account_line_list_action)d" 
--->
 
                         <!-- Registro -->
                         <button 
@@ -48,8 +39,6 @@
                                 string="Registro de Ventas" 
                         />
 
-
-
                         <!-- Contasis -->
                         <button 
                                 class="oe_highlight"
@@ -62,7 +51,6 @@
                                 string="Contasis"                               
                         />
 
-
                         <!-- Tarjetas -->
                         <button 
                                 class="oe_highlight"
@@ -70,21 +58,13 @@
                                 context="{
                                             'search_default_account_id': active_id, 
                                         }"
-                                
-                                name="%(payment_line_form_action)d" 
-                                
+                                name="%(payment_line_form_action)d"                                 
                                 string="Tarjetas"                               
                         />
-
                     </header>
-
-
-
-
 
                     <!-- Sheet -->  
                     <sheet>
-
                         <group  string="Total"  >
                             <group>
                                 <field name="name"/>
@@ -97,45 +77,24 @@
                                 <field name="total_count"/>
                             </group>
                         </group>
-<!--
-                                <field name="date"/>
-                                <field name="test_target"/>
-                                <field name="x_type"/>
--->
-
-
-
-
 
                         <!-- Notebook --> 
                         <notebook>  
                             <page string="Ventas">
-
-
                                 <field name="account_line">
                                     <tree string="Ventas"   limit="1000">
-
-
                                         <field name="serial_nr"/>
                                         <field name="state"/>
                                         <field name="x_type"/>
-
-
                                         <field name="date_time"/>
                                         <field name="date_char"/>
                                         <field name="time_char"/>
-
-
                                         <field name="patient"/>
                                         <field name="document"/>
                                         <field name="document_type"/>
-
-
                                         <field name="qty"/>
                                         <field name="product"/>
                                         <field name="product_type"/>
-                                        
-
                                         <field name="amount_net"/>                                      
                                         <field name="amount_tax"/>
                                         <field name="amount"/>
@@ -158,39 +117,24 @@
 -->
 
 
-
-
-
 <!-- TREE --> 
         <record model="ir.ui.view" id="account_contasis_view_order_tree">
             <field name="name">openhealth.account.contasis.tree</field>
             <field name="model">openhealth.account.contasis</field>
             <field name="arch" type="xml">
-
                 <tree string="Account Contasis Tree">
-                    
-
                     <field name="name"/>
-
                     <field name="date_begin"    string="Fecha Inicio"/>
-
                     <field name="date_end"      string="Fecha Fin"/>
-
                     <field name="owner"/>
-
                     <field name="x_type"/>
-
                     <field name="total_amount"/>
-
                     <field name="total_count"/>
                 </tree>
             </field>
         </record>
 <!--
-                    <field name="test_target"/>
                     <field name="date"          string="Fecha"  />
 -->
-
-
     </data>
 </openerp>
